media: xilinx: vtc: Fix warning for OVERFLOW_BEFORE_WIDEN
This patch fixes the warning for OVERFLOW_BEFORE_WIDEN.

The actual warning is -
For xvtc_generator_start():
Event overflow_before_widen: 	Potentially overflowing expression
"config->fps * config->hsize" with type "unsigned int" (32 bits,
unsigned) is evaluated using 32-bit arithmetic, and then used in a
context that expects an expression of type "unsigned long" (64 bits,
unsigned).
Event remediation: 	To avoid overflow, cast either "config->fps" or
"config->hsize" to type "unsigned long".

State: pending

+++ drivers/media/platform/xilinx/xilinx-vtc.c
@@ -185,7 +185,7 @@ int xvtc_generator_start(struct xvtc_device *xvtc,
 	if (!xvtc->has_generator)
 		return -ENXIO;
 
-	s_rate = config->fps * config->hsize * config->vsize;
+	s_rate = (unsigned long)config->fps * config->hsize * config->vsize;
 	ret = clk_set_rate(xvtc->xvip.clk, s_rate);
 	if (ret < 0)
 		return ret;
